Change logger to be a singleton import. Change config to be an import.
Use logger import in serverPlugins.
