+++ packages/gluestick/src/config.js
@@ -1,0 +1,4 @@
+import type { GSConfig } from './types';
+
+const config: GSConfig = JSON.parse(process.argv[2] || '{}');
+module.exports = config;
